Add implementation of {Get,GetAll,Set}() to dbus.service.object.Object
[work-in-progress]

This is a direct port of python3 application code. Get/Set/GetAll are
added directly to each Object. This is suboptimal but the idea is to get
comments and fix this.

The implementation uses "interface property map" which is basically
_dbus_class_table but I wrote that before realizing this. It will be
removed (along with all the support code) and everything will use existing
functionality.

I'm not sure how to import dbus.decorators.property without changing all
the existing uses of @property. I'll fix this and get it to work (along
with how __all__ exports this)

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>